dameon/c_seccomp: Moved c_seccomp.c to c_seccomp/seccomp.c
This is the beginning of restructuring the seccomp module to separate
files. The main seccomp handling should be take palce in the
'c_seccomp/seccomp.c' file while each syscall emulation code should
be moved into separate emulation specific files, e.g., mknod emulation
will be moved to c_seccomp/mknod.c.
